added pagination for viewing posts
