+++ Frontend/src/CSS/Pagination.css
@@ -1,0 +1,58 @@
+/* Pagination bar container */
+.pagination {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 20px;
+    padding: 10px;
+    background-color: #f9f9f9;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+/* Pagination button */
+.pagination button {
+    background-color: #fff;
+    border: 2px solid #ccc;
+    border-radius: 4px;
+    margin: 0 5px;
+    padding: 8px 12px;
+    color: #333;
+    font-size: 1em;
+    cursor: pointer;
+    transition: background-color 0.3s, color 0.3s, box-shadow 0.3s;
+}
+
+/* Hover effect for pagination button */
+.pagination button:hover {
+    background-color: #007BFF;
+    color: #fff;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+/* Active pagination button */
+.pagination button.active {
+    background-color: #007BFF;
+    color: #fff;
+    border: 2px solid #007BFF;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+/* Disabled pagination button */
+.pagination button:disabled {
+    background-color: #eee;
+    color: #999;
+    border: 2px solid #ccc;
+    cursor: not-allowed;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+    .pagination {
+        flex-wrap: wrap;
+    }
+
+    .pagination button {
+        margin: 5px 0;
+    }
+}
